indexer: refactor layout, tests
