+++ lib/bcoin-browser.js
@@ -59,8 +59,8 @@ bcoin.Mnemonic = require('./hd/mnemonic');
 bcoin.indexer = require('./indexer');
 bcoin.Indexer = require('./indexer/indexer');
 bcoin.ChainClient = require('./indexer/chainclient');
-bcoin.TXIndexer = require('./indexer/txindexer/txindexer');
-bcoin.AddrIndexer = require('./indexer/addrindexer/addrindexer');
+bcoin.TXIndexer = require('./indexer/txindexer');
+bcoin.AddrIndexer = require('./indexer/addrindexer');
 
 // Mempool
 bcoin.mempool = require('./mempool');
