+++ lib/bcoin.js
@@ -79,8 +79,8 @@ bcoin.define('Mnemonic', './hd/mnemonic');
 bcoin.define('indexer', './indexer');
 bcoin.define('Indexer', './indexer/indexer');
 bcoin.define('ChainClient', './indexer/chainclient');
-bcoin.define('TXIndexer', './indexer/txindexer/txindexer');
-bcoin.define('AddrIndexer', './indexer/addrindexer/addrindexer');
+bcoin.define('TXIndexer', './indexer/txindexer');
+bcoin.define('AddrIndexer', './indexer/addrindexer');
 
 // Mempool
 bcoin.define('mempool', './mempool');
