+++ lib/indexer/addrindexer.js
@@ -9,8 +9,19 @@
 const assert = require('assert');
 const bdb = require('bdb');
 const layout = require('./layout');
-const Address = require('../../primitives/address');
-const Indexer = require('../indexer');
+const Address = require('../primitives/address');
+const Indexer = require('./indexer');
+
+/*
+ * AddrIndexer Database Layout:
+ *  T[addr-hash][hash] -> dummy (tx by address)
+ *  C[addr-hash][hash][index] -> dummy (coin by address)
+*/
+
+Object.assign(layout, {
+  T: bdb.key('T', ['hash', 'hash256']),
+  C: bdb.key('C', ['hash', 'hash256', 'uint32'])
+});
 
 class AddrIndexer extends Indexer {
   /**
