+++ lib/indexer/chainclient.js
@@ -8,6 +8,7 @@
 
 const assert = require('assert');
 const AsyncEmitter = require('bevent');
+const Chain = require('../blockchain/chain');
 
 /**
  * Chain Client
@@ -22,6 +23,8 @@ class ChainClient extends AsyncEmitter {
 
   constructor(chain) {
     super();
+
+    assert(chain instanceof Chain);
 
     this.chain = chain;
     this.network = chain.network;
