+++ lib/indexer/index.js
@@ -7,6 +7,6 @@
 'use strict';
 
 exports.Indexer = require('./indexer');
-exports.TXIndexer = require('./txindexer/txindexer');
-exports.AddrIndexer = require('./addrindexer/addrindexer');
+exports.TXIndexer = require('./txindexer');
+exports.AddrIndexer = require('./addrindexer');
 exports.ChainClient = require('./chainclient');
