+++ lib/indexer/layout.js
@@ -10,6 +10,7 @@ const bdb = require('bdb');
 
 /*
  * Index Database Layout:
+ * To be extended by indexer implementations
  *  V -> db version
  *  O -> flags
  *  h[height] -> recent block hash
