+++ lib/indexer/txindexer.js
@@ -8,8 +8,17 @@
 
 const bdb = require('bdb');
 const layout = require('./layout');
-const TXMeta = require('../../primitives/txmeta');
-const Indexer = require('../indexer');
+const TXMeta = require('../primitives/txmeta');
+const Indexer = require('./indexer');
+
+/*
+ * TXIndexer Database Layout:
+ *  t[hash] -> extended tx
+*/
+
+Object.assign(layout, {
+  t: bdb.key('t', ['hash256'])
+});
 
 class TXIndexer extends Indexer {
   /**
