+++ lib/node/fullnode.js
@@ -16,8 +16,8 @@ const Miner = require('../mining/miner');
 const Node = require('./node');
 const HTTP = require('./http');
 const RPC = require('./rpc');
-const TXIndexer = require('../indexer/txindexer/txindexer');
-const AddrIndexer = require('../indexer/addrindexer/addrindexer');
+const TXIndexer = require('../indexer/txindexer');
+const AddrIndexer = require('../indexer/addrindexer');
 
 /**
  * Full Node
